fix(hub): close the review findings on the Android 15 QPR gate
The STAGE3 gate could pass both halves on the same numbering: nothing
verified that the GSI servicemanager actually speaks the r6+ protocol.
The interop binary now takes the expected numbering (`pre-r6`/`r6+`) and
fails unless the probe picked it, and the script passes it on each half,
replacing the log-grep cross-check. The script's pgrep pattern also never
matched the real servicemanager (relative argv[0]) while false-matching
the adb wrapper shell; match the unique cmdline with a bracket pattern.

Move the transaction-code pin tests out of the target_os-gated module
into a host-runnable one so every `cargo test` runs them, fix the
`isDeclared` code label (6/7, not 7/8), route gate-visible messages to
stdout, and condense the multi-paragraph docs on private items into the
public `android_15` module doc per the comment policy.
