+++ rsbinder/src/hub/servicemanager_15.rs
@@ -7,7 +7,9 @@
 //! the numbering split exists, and why nothing here parses the `Service`
 //! union — is on [`hub::android_15`](super::android_15), the module these
 //! items are re-exported through. This one is private, so its own docs would
-//! not ship.
+//! not ship. The transaction-code pins live in `hub/numbering_pins.rs`,
+//! which runs on every host `cargo test` (this module never does — it is
+//! behind `target_os = "android"`).
 
 include!(concat!(env!("OUT_DIR"), "/service_manager_15.rs"));
 
@@ -82,102 +84,5 @@ crate::hub::impl_sm_module_body! { @custom_check_service
                 None
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    //! Pin the transaction codes this module sends. They are the whole
-    //! reason it exists: `android-15.0.0_r6` shifted them and nothing in the
-    //! interface records that, so a regenerate from the wrong `.aidl` — or a
-    //! reordered one — would be silent everywhere else. On a real
-    //! `r6`+ service manager the wrong codes reach real methods, so the
-    //! failure is a `BAD_PARCELABLE` that names nothing.
-    use super::android::os::IServiceManager::transactions;
-    use crate::FIRST_CALL_TRANSACTION;
-
-    #[test]
-    fn transaction_codes_are_the_r6_numbering() {
-        // AOSP `android-15.0.0_r6`..`r36`, `libs/binder/aidl/android/os/
-        // IServiceManager.aidl` — declaration order is the numbering.
-        assert_eq!(transactions::r#getService, FIRST_CALL_TRANSACTION);
-        assert_eq!(transactions::r#getService2, FIRST_CALL_TRANSACTION + 1);
-        assert_eq!(transactions::r#checkService, FIRST_CALL_TRANSACTION + 2);
-        assert_eq!(transactions::r#addService, FIRST_CALL_TRANSACTION + 3);
-        assert_eq!(transactions::r#listServices, FIRST_CALL_TRANSACTION + 4);
-        assert_eq!(
-            transactions::r#registerForNotifications,
-            FIRST_CALL_TRANSACTION + 5
-        );
-        assert_eq!(
-            transactions::r#unregisterForNotifications,
-            FIRST_CALL_TRANSACTION + 6
-        );
-        assert_eq!(transactions::r#isDeclared, FIRST_CALL_TRANSACTION + 7);
-        assert_eq!(
-            transactions::r#getDeclaredInstances,
-            FIRST_CALL_TRANSACTION + 8
-        );
-        assert_eq!(transactions::r#updatableViaApex, FIRST_CALL_TRANSACTION + 9);
-        assert_eq!(
-            transactions::r#getUpdatableNames,
-            FIRST_CALL_TRANSACTION + 10
-        );
-        assert_eq!(
-            transactions::r#getConnectionInfo,
-            FIRST_CALL_TRANSACTION + 11
-        );
-        assert_eq!(
-            transactions::r#registerClientCallback,
-            FIRST_CALL_TRANSACTION + 12
-        );
-        assert_eq!(
-            transactions::r#tryUnregisterService,
-            FIRST_CALL_TRANSACTION + 13
-        );
-        assert_eq!(
-            transactions::r#getServiceDebugInfo,
-            FIRST_CALL_TRANSACTION + 14
-        );
-    }
-
-    /// The probe in [`hub::default`](crate::hub::default) sends code 14 and
-    /// reads the answer as "which interface is this". That only decides
-    /// anything while 14 is the last method here *and* one past the last
-    /// method of the Android 14 interface — the two facts this asserts.
-    #[cfg(feature = "android_14")]
-    #[test]
-    fn probe_code_14_separates_the_two_interfaces() {
-        use crate::hub::android_14::android::os::IServiceManager::transactions as prev;
-
-        assert_eq!(
-            transactions::r#getServiceDebugInfo,
-            FIRST_CALL_TRANSACTION + 14,
-            "code 14 must be a method here, or the probe reads `UnknownTransaction` \
-             on both interfaces"
-        );
-        assert_eq!(
-            prev::r#getServiceDebugInfo,
-            FIRST_CALL_TRANSACTION + 13,
-            "code 14 must be past the end of the Android 14 interface, or the probe \
-             gets an answer on both"
-        );
-
-        // Every method the two share, from `checkService` on, moved up
-        // exactly one. This is what makes the wrong module fail *late* and
-        // opaquely rather than not at all.
-        assert_eq!(transactions::r#checkService, prev::r#checkService + 1);
-        assert_eq!(transactions::r#addService, prev::r#addService + 1);
-        assert_eq!(
-            transactions::r#registerClientCallback,
-            prev::r#registerClientCallback + 1
-        );
-        assert_eq!(
-            transactions::r#tryUnregisterService,
-            prev::r#tryUnregisterService + 1
-        );
-        // `getService` is the one that did not move — the reason this module
-        // routes every lookup through it.
-        assert_eq!(transactions::r#getService, prev::r#getService);
     }
 }
